Day 7: Part 2: Working Solution
This is a pretty complete restructure of how I originally solved the problem.
Passing the entire tower around was tedious and a class started making more
sense, but in doing this I appear to have broken my tests.

I put a TODO in to come back to the test, but I did finish solving the problem.
Will fix before push.